Moved `eol` command to admin cog
This will hide it from `help` output
