+++ cogs/public.py
@@ -9,29 +9,12 @@ import pydest
 import cogs.utils.autocomplete
 from cogs.utils.autocomplete import metric_picker
 from cogs.utils.converters import locale_2_lang
-from cogs.utils.views import EOLButtons
 
 
 class Public(commands.Cog):
 
     def __init__(self, bot):
         self.bot = bot
-
-    @commands.command(
-        aliases=['top', 'support', 'online', 'lfglist', 'lfg', 'editlfg', 'setprefix', 'prefix', 'setclan', 'setlang'
-                 'rmnotifier', 'regnotifier', 'lfgcleanup']
-    )
-    async def eol(self, ctx, *args):
-        lang = 'en'
-        if ctx.guild is not None:
-            lang = ctx.bot.guild_lang(ctx.message.guild.id)
-        translations = ctx.bot.translations[lang]['msg']
-        response_embed = discord.Embed(title=translations['eol_title'],
-                                       description=translations['eol'].format(ctx.invoked_with,
-                                                                              translations['invite_btn']))
-        view = EOLButtons(ctx.bot.application_id, support_label=translations['support_btn'],
-                          invite_label=translations['invite_btn'])
-        await ctx.message.reply(embed=response_embed, view=view, mention_author=False)
 
     @commands.slash_command(
         name='top',
